requirements.txt
streamlit==1.35.0
pandas==2.2.2
requests==2.31.0
plotly==5.21.0
python-dotenv==1.0.0